helloWebApp/helloWebApp/src/main/java/io/github/shahroozsabet/HelloServlet.java
package io.github.shahroozsabet;

import jakarta.servlet.ServletException;
import jakarta.servlet.annotation.WebServlet;
import jakarta.servlet.http.HttpServlet;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;

import java.io.IOException;

/**
 * Created by Shahrooz on 8/24/2022.
 */
@WebServlet("/hello")
public class HelloServlet extends HttpServlet {

    protected void doPost(HttpServletRequest request, HttpServletResponse response)
            throws ServletException, IOException {
        String username = request.getParameter("username");
        request.setAttribute("username", "Hello " + username);
        request.getRequestDispatcher("response.jsp").forward(request, response);
    }
}
